requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.94.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.6
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.26.0.post1
typing_extensions==4.5.0
uvicorn==0.21.0
